requirements-dev.txt
flake8>=3.7
flake8-docstrings>=1.3
pytest>=4.2
pytest-aiohttp>=0.3
